requirements.txt
fastapi
uvicorn
pandas
ccxt
pandas
numpy
pandas_ta
requests
python-dotenv
apscheduler
psycopg2-binary
pydantic